Remove some s_all output: remove some macros we no longer need. Also fix some gcc 4.5 warnings: ../src/api/cur_stat.c:217:2: warning: 'csize' may be used uninitialized in this function ../src/api/cur_stat.c:222:21: warning: 'key_format' may be used uninitialized in this function ../src/api/cur_stat.c:223:23: warning: 'value_format' may be used uninitialized in this function ../src/utilities/util_load.c:127:25: warning: format '%llu' expects type 'long long unsigned int', but argument 3 has type 'uint64_t' ../src/utilities/util_load.c:141:3: warning: format '%llu' expects type 'long long unsigned int', but argument 3 has type 'uint64_t'

+++ src/utilities/util_load.c
@@ -34,7 +34,7 @@ main(int argc, char *argv[])
 	WT_SESSION *session;
 	WT_CURSOR *cursor;
 	struct record_t record;
-	uint64_t insert_count;
+	unsigned long long insert_count;
 	int ch, debug, eof, ret, text_input, tret, verbose;
 	const char *tablename, *table_config, *home;
 	char cursor_config[100], datasink[100];
